requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
Werkzeug==2.3.7
openpyxl==3.1.2
Pillow==10.0.1
gunicorn==21.2.0
python-dotenv==1.0.0
Flask-CORS==4.0.0
Flask-Session==0.5.0